Move BufferedOutputFiles class from split_icell8_fastqs.py to utils.
